ce-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ce-pipeline"
version = "0.1.0"
description = "Store-based Chunking + Embedding pipeline for RAG (chunks + vector index + BM25 index)."
readme = "README_EN.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Your Name"}]
dependencies = [
  "pyyaml>=6.0.1",
  "orjson>=3.10.0",
  "tqdm>=4.66.0",
  "regex>=2024.0.0",
  "python-dateutil>=2.9.0.post0",
  "rapidfuzz>=3.9.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "rank-bm25>=0.2.2",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12.20240808",
]

[project.scripts]
ce = "ce_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
